+++ test/helpers/TestSetup.sol
@@ -21,7 +21,6 @@ import {TestConstants} from "./TestConstants.sol";
  *         consistent setup and providing shared utilities across the test suite
  */
 abstract contract TestSetup is Test, TestConstants {
-    
     // Core contract instances - these will be deployed in each test
     CreatorRegistry public creatorRegistry;
     ContentRegistry public contentRegistry;
@@ -29,12 +28,12 @@ abstract contract TestSetup is Test, TestConstants {
     SubscriptionManager public subscriptionManager;
     CommerceProtocolIntegration public commerceIntegration;
     PriceOracle public priceOracle;
-    
+
     // Mock contracts for external dependencies
     MockERC20 public mockUSDC;
     MockCommerceProtocol public mockCommerceProtocol;
     MockQuoterV2 public mockQuoter;
-    
+
     // Test user addresses - using consistent addresses across all tests
     address public creator1 = address(0x1001);
     address public creator2 = address(0x1002);
@@ -43,15 +42,31 @@ abstract contract TestSetup is Test, TestConstants {
     address public admin = address(0x3001);
     address public feeRecipient = address(0x3002);
     address public operatorSigner = address(0x3003);
-    
+
     // Test constants for pricing and configuration
-    
+
     // Events for testing - we'll check these are emitted correctly
     event CreatorRegistered(address indexed creator, uint256 subscriptionPrice, uint256 timestamp, string profileData);
-    event ContentRegistered(uint256 indexed contentId, address indexed creator, string ipfsHash, string title, ContentRegistry.ContentCategory category, uint256 payPerViewPrice, uint256 timestamp);
+    event ContentRegistered(
+        uint256 indexed contentId,
+        address indexed creator,
+        string ipfsHash,
+        string title,
+        ContentRegistry.ContentCategory category,
+        uint256 payPerViewPrice,
+        uint256 timestamp
+    );
     event ContentPurchased(uint256 indexed contentId, address indexed buyer, uint256 price, uint256 timestamp);
-    event Subscribed(address indexed user, address indexed creator, uint256 price, uint256 platformFee, uint256 creatorEarning, uint256 startTime, uint256 endTime);
-    
+    event Subscribed(
+        address indexed user,
+        address indexed creator,
+        uint256 price,
+        uint256 platformFee,
+        uint256 creatorEarning,
+        uint256 startTime,
+        uint256 endTime
+    );
+
     /**
      * @dev Sets up the complete test environment
      * @notice This function is called before each test to ensure a clean state
@@ -60,22 +75,22 @@ abstract contract TestSetup is Test, TestConstants {
     function setUp() public virtual {
         // Set up the test environment with proper admin permissions
         vm.startPrank(admin);
-        
+
         // Deploy mock external dependencies first
         _deployMockDependencies();
-        
+
         // Deploy our core contracts with proper constructor parameters
         _deployCoreContracts();
-        
+
         // Configure contracts with proper roles and permissions
         _configureContracts();
-        
+
         // Set up test users with initial balances and permissions
         _setupTestUsers();
-        
+
         vm.stopPrank();
     }
-    
+
     /**
      * @dev Deploys mock contracts for external dependencies
      * @notice We use mocks to isolate our contract logic from external systems
@@ -84,19 +99,19 @@ abstract contract TestSetup is Test, TestConstants {
         // Deploy mock USDC with proper decimals and initial supply
         mockUSDC = new MockERC20("Mock USDC", "USDC", 6);
         mockUSDC.mint(admin, 1000000e6); // 1M USDC for testing
-        
+
         // Deploy mock Commerce Protocol for payment testing
         mockCommerceProtocol = new MockCommerceProtocol();
-        
+
         // Deploy mock Uniswap Quoter for price oracle testing
         mockQuoter = new MockQuoterV2();
-        
+
         console.log("Mock dependencies deployed:");
         console.log("- Mock USDC:", address(mockUSDC));
         console.log("- Mock Commerce Protocol:", address(mockCommerceProtocol));
         console.log("- Mock Quoter:", address(mockQuoter));
     }
-    
+
     /**
      * @dev Deploys all core platform contracts
      * @notice Order matters here due to constructor dependencies
@@ -104,28 +119,21 @@ abstract contract TestSetup is Test, TestConstants {
     function _deployCoreContracts() internal {
         // Deploy PriceOracle first (no dependencies)
         priceOracle = new PriceOracle();
-        
+
         // Deploy CreatorRegistry (depends on USDC)
         creatorRegistry = new CreatorRegistry(feeRecipient, address(mockUSDC));
-        
+
         // Deploy ContentRegistry (depends on CreatorRegistry)
         contentRegistry = new ContentRegistry(address(creatorRegistry));
-        
+
         // Deploy PayPerView (depends on multiple contracts)
-        payPerView = new PayPerView(
-            address(creatorRegistry),
-            address(contentRegistry),
-            address(priceOracle),
-            address(mockUSDC)
-        );
-        
+        payPerView =
+            new PayPerView(address(creatorRegistry), address(contentRegistry), address(priceOracle), address(mockUSDC));
+
         // Deploy SubscriptionManager (depends on multiple contracts)
-        subscriptionManager = new SubscriptionManager(
-            address(creatorRegistry),
-            address(contentRegistry),
-            address(mockUSDC)
-        );
-        
+        subscriptionManager =
+            new SubscriptionManager(address(creatorRegistry), address(contentRegistry), address(mockUSDC));
+
         // Deploy CommerceProtocolIntegration (depends on all contracts)
         commerceIntegration = new CommerceProtocolIntegration(
             address(mockCommerceProtocol),
@@ -135,7 +143,7 @@ abstract contract TestSetup is Test, TestConstants {
             feeRecipient,
             operatorSigner
         );
-        
+
         console.log("Core contracts deployed:");
         console.log("- CreatorRegistry:", address(creatorRegistry));
         console.log("- ContentRegistry:", address(contentRegistry));
@@ -143,7 +151,7 @@ abstract contract TestSetup is Test, TestConstants {
         console.log("- SubscriptionManager:", address(subscriptionManager));
         console.log("- CommerceIntegration:", address(commerceIntegration));
     }
-    
+
     /**
      * @dev Configures contracts with proper roles and integrations
      * @notice This sets up the permission structure that allows contracts to interact
@@ -153,23 +161,23 @@ abstract contract TestSetup is Test, TestConstants {
         creatorRegistry.grantPlatformRole(address(payPerView));
         creatorRegistry.grantPlatformRole(address(subscriptionManager));
         creatorRegistry.grantPlatformRole(address(commerceIntegration));
-        
+
         // Grant purchase recorder role to PayPerView in ContentRegistry
         contentRegistry.grantPurchaseRecorderRole(address(payPerView));
-        
+
         // Grant payment processor role to CommerceIntegration in PayPerView
         payPerView.grantPaymentProcessorRole(address(commerceIntegration));
-        
+
         // Grant subscription processor role to CommerceIntegration in SubscriptionManager
         subscriptionManager.grantSubscriptionProcessorRole(address(commerceIntegration));
-        
+
         // Set up CommerceIntegration with other contract addresses
         commerceIntegration.setPayPerView(address(payPerView));
         commerceIntegration.setSubscriptionManager(address(subscriptionManager));
-        
+
         console.log("Contract permissions configured");
     }
-    
+
     /**
      * @dev Sets up test users with initial USDC balances
      * @notice This ensures all test users have sufficient funds for testing
@@ -180,16 +188,16 @@ abstract contract TestSetup is Test, TestConstants {
         users[1] = creator2;
         users[2] = user1;
         users[3] = user2;
-        
+
         // Give each user 1000 USDC for testing
-        for (uint i = 0; i < users.length; i++) {
+        for (uint256 i = 0; i < users.length; i++) {
             mockUSDC.mint(users[i], 1000e6);
             vm.deal(users[i], 10 ether); // Also give them ETH for gas
         }
-        
+
         console.log("Test users set up with initial balances");
     }
-    
+
     /**
      * @dev Helper function to register a creator with default settings
      * @param creator The address to register as creator
@@ -198,7 +206,7 @@ abstract contract TestSetup is Test, TestConstants {
     function registerCreator(address creator) internal returns (bool success) {
         return registerCreator(creator, DEFAULT_SUBSCRIPTION_PRICE, "Default profile");
     }
-    
+
     /**
      * @dev Helper function to register a creator with custom settings
      * @param creator The address to register as creator
@@ -206,7 +214,10 @@ abstract contract TestSetup is Test, TestConstants {
      * @param profileData IPFS hash for profile data
      * @return success Whether the registration was successful
      */
-    function registerCreator(address creator, uint256 subscriptionPrice, string memory profileData) internal returns (bool success) {
+    function registerCreator(address creator, uint256 subscriptionPrice, string memory profileData)
+        internal
+        returns (bool success)
+    {
         vm.startPrank(creator);
         try creatorRegistry.registerCreator(subscriptionPrice, profileData) {
             success = true;
@@ -216,7 +227,7 @@ abstract contract TestSetup is Test, TestConstants {
         vm.stopPrank();
         return success;
     }
-    
+
     /**
      * @dev Helper function to register content with default settings
      * @param creator The creator address
@@ -225,7 +236,7 @@ abstract contract TestSetup is Test, TestConstants {
     function registerContent(address creator) internal returns (uint256 contentId) {
         return registerContent(creator, DEFAULT_CONTENT_PRICE, "Sample content");
     }
-    
+
     /**
      * @dev Helper function to register content with custom settings
      * @param creator The creator address
@@ -233,26 +244,24 @@ abstract contract TestSetup is Test, TestConstants {
      * @param title The content title
      * @return contentId The ID of the registered content
      */
-    function registerContent(address creator, uint256 price, string memory title) internal returns (uint256 contentId) {
+    function registerContent(address creator, uint256 price, string memory title)
+        internal
+        returns (uint256 contentId)
+    {
         vm.startPrank(creator);
-        
+
         string[] memory tags = new string[](2);
         tags[0] = "test";
         tags[1] = "content";
-        
+
         contentId = contentRegistry.registerContent(
-            "QmTestHash123456789",
-            title,
-            "Test description",
-            ContentRegistry.ContentCategory.Article,
-            price,
-            tags
+            "QmTestHash123456789", title, "Test description", ContentRegistry.ContentCategory.Article, price, tags
         );
-        
+
         vm.stopPrank();
         return contentId;
     }
-    
+
     /**
      * @dev Helper function to approve USDC spending
      * @param user The user address
@@ -263,7 +272,7 @@ abstract contract TestSetup is Test, TestConstants {
         vm.prank(user);
         mockUSDC.approve(spender, amount);
     }
-    
+
     /**
      * @dev Helper function to check if two strings are equal
      * @param a First string
@@ -273,7 +282,7 @@ abstract contract TestSetup is Test, TestConstants {
     function stringEqual(string memory a, string memory b) internal pure returns (bool equal) {
         return keccak256(bytes(a)) == keccak256(bytes(b));
     }
-    
+
     /**
      * @dev Helper function to advance time in tests
      * @param timeToAdvance The time to advance in seconds
@@ -281,7 +290,7 @@ abstract contract TestSetup is Test, TestConstants {
     function advanceTime(uint256 timeToAdvance) internal {
         vm.warp(block.timestamp + timeToAdvance);
     }
-    
+
     /**
      * @dev Helper function to expect a specific revert message
      * @param expectedRevert The expected revert message
